fix conversion warnings

+++ r/src/array.c
@@ -55,7 +55,7 @@ SEXP nanoarrow_c_array_set_length(SEXP array_xptr, SEXP length_sexp) {
     Rf_error("array$length must be finite and greater than zero");
   }
 
-  array->length = length;
+  array->length = (int64_t)length;
   return R_NilValue;
 }
 
@@ -70,7 +70,7 @@ SEXP nanoarrow_c_array_set_null_count(SEXP array_xptr, SEXP null_count_sexp) {
     Rf_error("array$null_count must be finite and greater than -1");
   }
 
-  array->null_count = null_count;
+  array->null_count = (int64_t)null_count;
   return R_NilValue;
 }
 
@@ -85,7 +85,7 @@ SEXP nanoarrow_c_array_set_offset(SEXP array_xptr, SEXP offset_sexp) {
     Rf_error("array$offset must be finite and greater than zero");
   }
 
-  array->offset = offset;
+  array->offset = (int64_t)offset;
   return R_NilValue;
 }
