+++ r/src/materialize_chr.h
@@ -79,7 +79,7 @@ static inline int nanoarrow_materialize_chr(struct RConverter* converter) {
     } else {
       item = ArrowArrayViewGetStringUnsafe(src->array_view, src->offset + i);
       SET_STRING_ELT(dst->vec_sexp, dst->offset + i,
-                     Rf_mkCharLenCE(item.data, item.size_bytes, CE_UTF8));
+                     Rf_mkCharLenCE(item.data, (int)item.size_bytes, CE_UTF8));
     }
   }
